pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Automated job search pipeline — scrapes, scores, and generates tailored applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openai>=1.82.0",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
